refactor: split transcript row support modules
Separate the remaining transcript row support code by responsibility. Transcript item init structs, DB/display conversion, and tool-burst aggregation now live in transcript_item_init.rs, while shared GTK rendering helpers live in transcript_row_rendering.rs.

This removes the misleading transcript_row.rs module name after the legacy row factory was deleted. The typed ListView path now imports data preparation and rendering helpers from modules that describe what they actually own.

Verification:

- cargo fmt --all -- --check

- cargo clippy --all -- -D warnings

- cargo test transcript --lib --no-fail-fast

- cargo test typed_transcript_row --lib --no-fail-fast

- jscpd --reporters ai --format rust --min-lines 8 --min-tokens 80 src/ui/transcript_item_init.rs src/ui/transcript_row_rendering.rs src/ui/typed_transcript_row.rs src/ui/transcript_item_data.rs
