+++ src/ui/mod.rs
@@ -16,5 +16,6 @@ pub mod tool_preview;
 pub mod tool_renderers;
 pub mod transcript_display;
 pub mod transcript_item_data;
-pub mod transcript_row;
+pub mod transcript_item_init;
+pub mod transcript_row_rendering;
 pub mod typed_transcript_row;
